Fix: earningRate 미반영 오류 수정중

See merge request s09-fintech-finance-sub2/S09P22B109!115

+++ backend/economius/src/main/java/com/ssafy/economius/game/entity/redis/PortfolioBuildings.java
@@ -81,7 +81,7 @@ public class PortfolioBuildings {
     private void calculateEarnings(int totalBoughtPrice, int totalCurrentPrice) {
         int gap = totalCurrentPrice - totalBoughtPrice;
         int newEarningRate = 0;
-        if (totalBoughtPrice != 0) newEarningRate = (gap / totalBoughtPrice) * 100;
+        if (totalBoughtPrice != 0) newEarningRate = (int)((gap / (float)totalBoughtPrice) * 100);
         log.info("부동산 earning price 계산중 -> totalBoughtPrice: " + totalBoughtPrice
                 + ", totalCurrentPrice: " + totalCurrentPrice
                 + ", gap: " + gap
